BE/requirements.txt
fastapi
uvicorn
google-generativeai
pydantic
python-multipart
deepface
opencv-python-headless
numpy
Pillow